+++ frontend/projects/public/src/app/components/home/home.component.spec.ts
@@ -62,6 +62,7 @@ describe('HomeComponent', () => {
               ownerName: 'Mock Owner', ownerHeadline: 'Principal Software Engineer',
               ownerDescription: 'Desc.', socialLinks: [],
               analyticsId: '',
+                availability: 'listening',
             }),
           },
         }